Move enrollToGroupTraining business logic from GroupTrainingsDbRepository to UserGroupTrainingService.

+++ trainings/src/main/java/com/healthy/gym/trainings/service/group/training/UserGroupTrainingServiceImpl.java
@@ -1,10 +1,15 @@
 package com.healthy.gym.trainings.service.group.training;
 
+import com.healthy.gym.trainings.data.document.GroupTrainings;
+import com.healthy.gym.trainings.data.document.UserDocument;
 import com.healthy.gym.trainings.data.repository.GroupTrainingsDbRepositoryImpl;
-import com.healthy.gym.trainings.exception.training.TrainingEnrollmentException;
+import com.healthy.gym.trainings.data.repository.GroupTrainingsRepository;
+import com.healthy.gym.trainings.data.repository.UserDAO;
 import com.healthy.gym.trainings.exception.invalid.InvalidDateException;
 import com.healthy.gym.trainings.exception.invalid.InvalidHourException;
 import com.healthy.gym.trainings.exception.notexisting.NotExistingGroupTrainingException;
+import com.healthy.gym.trainings.exception.notfound.UserNotFoundException;
+import com.healthy.gym.trainings.exception.training.TrainingEnrollmentException;
 import com.healthy.gym.trainings.model.response.GroupTrainingPublicResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,10 +20,18 @@ import java.util.List;
 public class UserGroupTrainingServiceImpl implements UserGroupTrainingService {
 
     private final GroupTrainingsDbRepositoryImpl groupTrainingsDbRepositoryImpl;
+    private final GroupTrainingsRepository groupTrainingsRepository;
+    private final UserDAO userRepository;
 
     @Autowired
-    public UserGroupTrainingServiceImpl(GroupTrainingsDbRepositoryImpl groupTrainingsDbRepositoryImpl) {
+    public UserGroupTrainingServiceImpl(
+            GroupTrainingsDbRepositoryImpl groupTrainingsDbRepositoryImpl,
+            GroupTrainingsRepository groupTrainingsRepository,
+            UserDAO userRepository
+    ) {
         this.groupTrainingsDbRepositoryImpl = groupTrainingsDbRepositoryImpl;
+        this.groupTrainingsRepository = groupTrainingsRepository;
+        this.userRepository = userRepository;
     }
 
     @Override
@@ -29,14 +42,26 @@ public class UserGroupTrainingServiceImpl implements UserGroupTrainingService {
     }
 
     @Override
-    public void enrollToGroupTraining(String trainingId, String clientId) throws TrainingEnrollmentException {
+    public void enrollToGroupTraining(String trainingId, String clientId)
+            throws TrainingEnrollmentException, NotExistingGroupTrainingException, UserNotFoundException {
+
+        GroupTrainings groupTrainings = groupTrainingsRepository.findFirstByTrainingId(trainingId);
+        if (groupTrainings == null) throw new NotExistingGroupTrainingException();
+
+        UserDocument newParticipant = userRepository.findByUserId(clientId);
+        if (newParticipant == null) throw new UserNotFoundException();
+
         if (!groupTrainingsDbRepositoryImpl.isAbilityToGroupTrainingEnrollment(trainingId))
             throw new TrainingEnrollmentException("Cannot enroll to this training");
         if (groupTrainingsDbRepositoryImpl.isClientAlreadyEnrolledToGroupTraining(trainingId, clientId))
             throw new TrainingEnrollmentException("Client is already enrolled to this training");
 
-        groupTrainingsDbRepositoryImpl.enrollToGroupTraining(trainingId, clientId);
+        List<UserDocument> participants = groupTrainings.getParticipants();
+        participants.add(newParticipant);
+        groupTrainings.setParticipants(participants);
+        groupTrainingsRepository.save(groupTrainings);
 
+        // TODO nie rozumiem tego kodu poniżej, czemu on tutaj jest?
         if (groupTrainingsDbRepositoryImpl.isClientAlreadyExistInReserveList(trainingId, clientId))
             groupTrainingsDbRepositoryImpl.removeFromReserveList(trainingId, clientId);
     }
